Move fs integration tests
